+++ Tests/EncryptionWrapperTest.php
@@ -15,6 +15,7 @@ use Defuse\Crypto\Exception\CryptoException as BaseCryptoException;
 use Defuse\Crypto\Exception\EnvironmentIsBrokenException;
 use Mes\Security\CryptoBundle\EncryptionWrapper;
 use Mes\Security\CryptoBundle\Exception\CryptoException;
+use Symfony\Component\Filesystem\Filesystem;
 
 /**
  * Class EncryptionWrapperTest.
@@ -87,6 +88,71 @@ class EncryptionWrapperTest extends \PHPUnit_Framework_TestCase
         }
 
         $this->wrapper->decrypt('ThisIsACipherText', $this->getMock('Mes\Security\CryptoBundle\Model\KeyInterface'));
+    }
+
+    public function testEncryptFileEncryptsFile()
+    {
+        $this->encryption->expects($this->once())
+                         ->method('encryptFile')
+                         ->will($this->returnCallback(function ($input, $output) {
+                             $fs = new Filesystem();
+                             $fs->dumpFile($output, '');
+                         }));
+
+        $this->wrapper->encryptFile(__DIR__.'/file.txt', __DIR__.'/file.crypto', $this->getMock('Mes\Security\CryptoBundle\Model\KeyInterface'));
+
+        $this->assertFileExists(__DIR__.'/file.crypto');
+
+        unlink(__DIR__.'/file.crypto');
+    }
+
+    public function testDecryptFileDecryptsEncryptedFile()
+    {
+        $this->encryption->expects($this->once())
+                         ->method('decryptFile')
+                         ->will($this->returnCallback(function ($input, $output) {
+                             $fs = new Filesystem();
+                             $fs->dumpFile($output, 'Plain text');
+                         }));
+
+        $this->wrapper->decryptFile(__DIR__.'/file.crypto', __DIR__.'/file.txt', $this->getMock('Mes\Security\CryptoBundle\Model\KeyInterface'));
+
+        $this->assertFileExists(__DIR__.'/file.txt');
+        $this->assertContains('Plain text', file_get_contents(__DIR__.'/file.txt'));
+
+        unlink(__DIR__.'/file.txt');
+    }
+
+    /**
+     * @expectedException \Mes\Security\CryptoBundle\Exception\CryptoException
+     */
+    public function testEncryptFileThrowsException()
+    {
+        try {
+            $this->encryption->expects($this->once())
+                             ->method('encryptFile')
+                             ->will($this->throwException(new BaseCryptoException()));
+        } catch (BaseCryptoException $e) {
+            $this->throwException(new CryptoException());
+        }
+
+        $this->wrapper->encryptFile(__DIR__.'/file.txt', __DIR__.'/file.crypto', $this->getMock('Mes\Security\CryptoBundle\Model\KeyInterface'));
+    }
+
+    /**
+     * @expectedException \Mes\Security\CryptoBundle\Exception\CryptoException
+     */
+    public function testDecryptFileThrowsException()
+    {
+        try {
+            $this->encryption->expects($this->once())
+                             ->method('decryptFile')
+                             ->will($this->throwException(new BaseCryptoException()));
+        } catch (BaseCryptoException $e) {
+            $this->throwException(new CryptoException());
+        }
+
+        $this->wrapper->decryptFile(__DIR__.'/file.crypto', __DIR__.'/file.txt', $this->getMock('Mes\Security\CryptoBundle\Model\KeyInterface'));
     }
 
     protected function setUp()