+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 Changelog
 =========
 
+1.1.0 (2017-05-xx)
+------------------
+* Added ability to encrypt files:
+    - `EncryptionInterface::encryptFile`
+    - `EncryptionInterface::decryptFile`
+
 1.0.1 (2017-03-21)
 ------------------
 * Fixed: `KeyGeneratorCommand` shows process debug information (`-vvv`) only if `--dir` option is defined.